Fixes from renamings
